This removes the __construct() method, and moves the dependencies to the handle() method.
This is required in order to run artisan when there is no database connection
When dependencies are injected through the constructor, the artisan command cannot be run without a database connection.
Moving the dependencies to the handle() method ensures that the other artisan commands can be run, an only database specific commands like this cannot be run.