created an index page for businesses
I've created a new set of admin/ views in the views/admin folder

I've created a new BusinessController under the `Admin` namespace.

I've created a new Route::group() in the routes/web.php file where the url prefix is admin and the controller namespace is Admin.

You can see how I'm specifying the controller her instead of an anonymous function.

I've created an admin layout file to contain the layout for our admin views. It uses the `@yield('content')` line to allow us to extend it in the `index.blade.php` file and inject our own html content for that page.