added a couple of new routes, one for the `business/new` and one for the post request to `business`. Note how the one for `admin.business.store` is `Route::post()` not `Route::get()`.
Also note how I'm naming the routes, which I then use to create urls to those routes in the blade files or the controller (see the store method in the controller and the html for the form in the admin/business/new.blade.php template).

Also note how the redirect is happening in the store method of the controller.