So here I've added the show and update routes for a business.
I've added a get route and put route on `business/{id}`, which shows how to put a parameter into a url.

You can see how in Use the id parameter in the controller for the `show` and `update` methods.

I've had to add a new method to the BusinessRepository to find a single Business.

I've created a new `admin.business.show` template that is basically the same form as the `admin.business.new` but with values. (There is a way to do this where you can use the same template, but we can chat about that later).